rename screen controller subfolder entry file as index.js for clarity and also reduce text when importing
